fix: Separate RawApi & JsonApiDocumentCollection
